computations ducks file combined, adjusting references in other files
