+++ packages/coinstac-ui/app/render/state/root-reducer.js
@@ -1,7 +1,6 @@
 import { combineReducers } from 'redux';
 import auth from './ducks/auth';
-import computation from './ducks/computation';
-import computations from './ducks/computations';
+import computationsState from './ducks/computations';
 import consortia from './ducks/consortia';
 import consortiaPage from './ducks/consortia-page';
 import loading from './ducks/loading';
@@ -12,8 +11,7 @@ import { reducer as form } from 'redux-form';
 
 const rootReducer = combineReducers({
   auth,
-  computation,
-  computations,
+  computationsState,
   consortia,
   consortiaPage,
   form,
